test: move the AIProviderError field tests to test_errors.py
They exercise errors.py alone with no ai_client dependency, and test_errors.py
already hosts TestExceptionHierarchy for the sibling exception types.
